# openapi.yaml
openapi: 3.0.0
info:
  title: My Collection
  description: >-
    ### Welcome to Postman! This is your first collection.


    Collections are your starting point for building and testing APIs. You can
    use this one to:


    • Group related requests  

    • Test your API in real-world scenarios  

    • Document and share your requests


    Update the name and overview whenever you’re ready to make it yours.


    [Learn more about Postman
    Collections.](https://learning.postman.com/docs/collections/collections-overview/)
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/users/unban:
    patch:
      tags:
        - CSC309 Project > Users > Ban Users
      summary: Unban a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                userId: 2
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...